Add files via upload
This program implements a credit card validation system in C. It uses Luhn’s Algorithm to check whether a card number is valid and identifies the issuing company (VISA, AMEX, MASTERCARD) based on the number’s prefix and length. 
Features:
1) Calculates the checksum of the card number using Luhn’s Algorithm.
2)Rejects invalid card numbers.
3)Detects and prints the card type: VISA (prefix 4 only), AMEX, and MASTERCARD

Enter a credit card number when prompted, and the program will output the result.