Rename Project 0 CS50.sb3 to Project 0 CS50
You have been summoned for the royal fruit delivery. Collect fruits in time and save your head !

A simple interactive Scratch project where the player delivers different fruits. The project demonstrates basic event handling, motion, and condition checking in Scratch. It is designed as an introduction to game logic, using sprites, variables, and loops to create an engaging activity.
This was one of my first projects experimenting with Scratch to learn the fundamentals of programming and game design.